Support for compiling in windows.
